split report.rs to different files
