+++ crates/karva_coverage/src/report/json.rs
@@ -1,0 +1,75 @@
+use std::collections::BTreeMap;
+
+use anyhow::{Context, Result};
+use serde::Serialize;
+
+use super::shared::{FileRow, missing_lines, percent, totals_row};
+
+#[derive(Serialize)]
+struct JsonSummary {
+    covered_lines: u32,
+    num_statements: u32,
+    percent_covered: f64,
+    missing_lines: Vec<u32>,
+    excluded_lines: Vec<u32>,
+}
+
+#[derive(Serialize)]
+struct JsonFileReport {
+    executed_lines: Vec<u32>,
+    summary: JsonSummary,
+    missing_lines: Vec<u32>,
+    excluded_lines: Vec<u32>,
+}
+
+#[derive(Serialize)]
+struct JsonReport {
+    meta: JsonMeta,
+    files: BTreeMap<String, JsonFileReport>,
+    totals: JsonSummary,
+}
+
+#[derive(Serialize)]
+struct JsonMeta {
+    format: u32,
+    version: &'static str,
+}
+
+pub fn build_json_report(rows: &[FileRow]) -> Result<String> {
+    let files = rows
+        .iter()
+        .map(|row| {
+            (
+                row.name.clone(),
+                JsonFileReport {
+                    executed_lines: row.executed.clone(),
+                    summary: json_summary(row),
+                    missing_lines: missing_lines(row),
+                    excluded_lines: Vec::new(),
+                },
+            )
+        })
+        .collect();
+
+    let totals_row = totals_row(rows);
+    let report = JsonReport {
+        meta: JsonMeta {
+            format: 2,
+            version: "karva",
+        },
+        files,
+        totals: json_summary(&totals_row),
+    };
+
+    serde_json::to_string_pretty(&report).context("failed to serialize coverage json")
+}
+
+fn json_summary(row: &FileRow) -> JsonSummary {
+    JsonSummary {
+        covered_lines: row.hit,
+        num_statements: row.stmts,
+        percent_covered: percent(row.stmts, row.miss),
+        missing_lines: missing_lines(row),
+        excluded_lines: Vec::new(),
+    }
+}
